requirement.txt
streamlit
moviepy
openai-whisper
transformers
torch
